remove platform completely
