+++ packages/dd-trace/src/pkg.js
