+++ packages/dd-trace/test/pkg-loader.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const pkg = require('../../../src/platform/node/pkg.js')
+const pkg = require('../src/pkg.js')
 
 // eslint-disable-next-line no-console
 console.log(JSON.stringify(pkg))
